AP-9610: Added setting to automatically assign a palette to new values
Added combobox and setting to palettes panel; the user now has the
option to fail on new values in the table (as before) or to assign a
color of set1, set2 or set3 from the palettes panel. The assignment
does not change old values and assigns the least used colors per set.
Updated Node Description with text to new setting.

AP-9610 (Make Color Manager able to dynamically update possible values when using Palettes)